add some annotations to my entities like JsonIgnore and JsonManagedReference and JsonBackReference and delete some like Data and override ToString to avoid Stack overflow and Avoid data duplicate insertions by save them twice by presist cascade and manually by the repository and finish Category And product controllers

+++ src/main/java/API/com/example/E_COMMERCY/dto/orderItem/OrderItemMapper.java
@@ -18,7 +18,7 @@ public class OrderItemMapper {
     public OrderItemResponseDto toDto(OrderItem orderItem){
         return OrderItemResponseDto.builder()
                 .productResponseDto(productMapper.toDto(orderItem.getProduct()))
-                .id(orderItem.getId())
+                .id((long) orderItem.getId())
                 .Quantity(orderItem.getQuantity())
                 .build();
     }
